+++ rust/guest/k16-shared-kraft/src/main.rs
@@ -59,6 +59,18 @@ pub unsafe extern "C" fn stat(path: *const u8, len: usize, metadata: *mut u8) ->
 }
 
 #[no_mangle]
+pub unsafe extern "C" fn rename(request: *const u8, len: usize) -> u32 {
+    unsafe {
+        __k16_syscall3(
+            k16_abi::syscall::RENAME,
+            request as usize as u32,
+            len as u32,
+            0,
+        )
+    }
+}
+
+#[no_mangle]
 pub unsafe extern "C" fn mkdir(path: *const u8, len: usize) -> u32 {
     unsafe { __k16_syscall3(k16_abi::syscall::MKDIR, path as usize as u32, len as u32, 0) }
 }